+++ node/src/main/kotlin/net/corda/node/ArgsParser.kt
@@ -34,6 +34,8 @@ class ArgsParser {
     private val noLocalShellArg = optionParser.accepts("no-local-shell", "Do not start the embedded shell locally.")
     private val isRegistrationArg = optionParser.accepts("initial-registration", "Start initial node registration with Corda network to obtain certificate from the permissioning server.")
     private val isVersionArg = optionParser.accepts("version", "Print the version and exit")
+    private val justGenerateNodeInfoArg = optionParser.accepts("just-generate-node-info",
+            "Perform the node start-up task necessary to generate its nodeInfo, save it to disk, then quit")
     private val helpArg = optionParser.accepts("help").forHelp()
 
     fun parse(vararg args: String): CmdLineOptions {
@@ -50,7 +52,9 @@ class ArgsParser {
         val isVersion = optionSet.has(isVersionArg)
         val noLocalShell = optionSet.has(noLocalShellArg)
         val sshdServer = optionSet.has(sshdServerArg)
-        return CmdLineOptions(baseDirectory, configFile, help, loggingLevel, logToConsole, isRegistration, isVersion, noLocalShell, sshdServer)
+        val justGenerateNodeInfo = optionSet.has(justGenerateNodeInfoArg)
+        return CmdLineOptions(baseDirectory, configFile, help, loggingLevel, logToConsole, isRegistration, isVersion,
+                noLocalShell, sshdServer, justGenerateNodeInfo)
     }
 
     fun printHelp(sink: PrintStream) = optionParser.printHelpOn(sink)
@@ -64,7 +68,8 @@ data class CmdLineOptions(val baseDirectory: Path,
                           val isRegistration: Boolean,
                           val isVersion: Boolean,
                           val noLocalShell: Boolean,
-                          val sshdServer: Boolean) {
+                          val sshdServer: Boolean,
+                          val justGenerateNodeInfo : Boolean) {
     fun loadConfig() = ConfigHelper
                 .loadConfig(baseDirectory, configFile)
                 .parseAs<FullNodeConfiguration>()